requirements.txt
PyYAML>=6.0
Flask>=2.3.0
requests>=2.31.0
Pillow>=10.0.0